Clean up a collider's contact manifolds when it's removed from the scene
Also change collision layer variables to use short instead of int
(layers were always ultimately stored as short anyway).

+++ src/physics/collision_object.c
@@ -10,7 +10,7 @@
 #include "math/mathf.h"
 #include "mesh_collider.h"
 
-void collisionObjectInit(struct CollisionObject* object, struct ColliderTypeData *collider, struct RigidBody* body, float mass, int collisionLayers) {
+void collisionObjectInit(struct CollisionObject* object, struct ColliderTypeData *collider, struct RigidBody* body, float mass, short collisionLayers) {
     object->collider = collider;
     object->body = body;
     object->position = &body->transform.position;
@@ -24,7 +24,7 @@ void collisionObjectInit(struct CollisionObject* object, struct ColliderTypeData
     object->manifoldIds = 0;
 }
 
-void collisionObjectReInit(struct CollisionObject* object, struct ColliderTypeData *collider, struct RigidBody* body, float mass, int collisionLayers) {
+void collisionObjectReInit(struct CollisionObject* object, struct ColliderTypeData *collider, struct RigidBody* body, float mass, short collisionLayers) {
     object->collider = collider;
     object->body = body;
     collisionObjectUpdateBB(object);
