+++ src/physics/collision_scene.h
@@ -47,8 +47,8 @@ struct Transform* collisionSceneTransformToPortal(int fromPortal);
 void collisionScenePushObjectsOutOfPortal(int portalIndex);
 void collisionSceneCheckUnwokenObjectsNearPortal(int portalIndex);
 
-int collisionSceneRaycast(struct CollisionScene* scene, int roomIndex, struct Ray* ray, int collisionLayers, float maxDistance, int passThroughPortals, struct RaycastHit* hit);
-int collisionSceneRaycastOnlyDynamic(struct CollisionScene* scene, struct Ray* ray, int collisionLayers, float maxDistance, struct RaycastHit* hit);
+int collisionSceneRaycast(struct CollisionScene* scene, int roomIndex, struct Ray* ray, short collisionLayers, float maxDistance, int passThroughPortals, struct RaycastHit* hit);
+int collisionSceneRaycastOnlyDynamic(struct CollisionScene* scene, struct Ray* ray, short collisionLayers, float maxDistance, struct RaycastHit* hit);
 
 void collisionSceneGetPortalTransform(int fromPortal, struct Transform* out);
